Fixes (reference) MIR and updates documentation
Several functions were templated, but the implementations were left in the cpp file.
This moves them to the hpp file.
